refactor: CartItem의 디렉토리 구조를 개선한다
